pyproject.toml
[project]
name = "signaltwice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "polars>=1.38.0",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
